Add files via upload
that program is for control internet and CPU
internet:shows your Internet usage amount and you can creates a limit, and if the user exceeds that limit, it gives them a warning.
CPU:It shows your CPU and allows you to write three numbers. If the CPU exceeds any of these numbers, it gives an error sound. You can also choose that if it reaches the third error five times, it will shut down your computer